Move RunResult class into multiqc.py
